apps/backend/requirements.txt
fastapi
uvicorn
pydantic-settings
sqlalchemy
psycopg2-binary
pydantic[email]
passlib[bcrypt]
python-jose[cryptography]
alembic==1.16.4
PyJWT==2.10.1
psycopg2-binary==2.9.10
sqlmodel==0.0.24